Add unique option when generating score
The unique option means that in scoring each entity (author etc) can only use a paper once for scoring.
That is if an author in multiple papers reference a paper of the user, only one of the citations will be counted towards the score.
To deal with the case in-which the citations do not have the same influence score, the largest is taken (maximum influence counted).