Update browse tables and connect to flower page
Place browse lists with cached details in data tables on the browse page. Also add functionality to click on user row and be taken to their respective flower.
This involved making some changes to the backend functions such as ignoring saved and saving caches. Note: Using API data not elastic.