Change the scoring for create page searches to include citation count
Add CitationCount as a variable to the score function when searching for authors, journals, papers, conferences and affiliations.
This is particularly useful for non-author searches because it helps to identify the biggest entities that are similar in name to the search.
The formula used is _score^3 * log(CitationCount + 10), chosen to favour entities that closely match the search term.
From ad hoc testing, searching exact names for entities with low citation counts usually produces the correct entity in the first or second position,.