make default flower config a function to prevent being overwritten
Previously the flower config being imported from high_level_get_flower and
a the config variable was made a pointer to that dictionary. When values
of config were subsequently changed they would update the globally stored
default dictionary. This was causing issues such as incorrect number of petals
on fresh flowers.